Create conditional execution
`cond` (or `?`) is a prefix that will pop the top value of the stack.
If it's zero, then the operation will execute, otherwise it won't.
Note that it is fine to *try* to execute an illegal op, but not if
the op actually runs.

For example, `cond add` is fine if the stack is [0], but not fine if
the stack is [1] because the add will also pop 2 more values.

The quiet prefix only effects the op, not the prefix.

Working on #10