Create qcond
qcond is a prefix that checks if the top element is not zero.
If it is, it executes the instruction on the elements _below_ the top element.
Any pushes will be inserted under the top element.

You cannot use qcond and cond in the same instruction.
All the restictions to cond also apply to qcond (no empty stack, op must be valid if it executes, etc)

Resolves #10