Check instructions for well-formedness
This commit adds `is_well_formed`, which checks if a particular instruction is
well formed or not. This is defined as either a naked label or an instruction of the form

<prefixes> [op] <args or label>

It does not care if the actual instruction does not work (for example, 'push @label')
nor does it typecheck anything. It only cares about syntax.

This commit also adds a bunch of small helper functions.

Working on #33